+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Changed
+- Add loadError if "go to" field cannot be found.
+- Fire "gotohidden.enketo" event if "go to" field is hidden.
+
 [4.35.0] - 2017-05-26
 ---------------------
 ##### Added